+++ packages/types/src/create/createType.ts
@@ -12,18 +12,25 @@ import { createClass } from './createClass';
 
 // With isPedantic, actually check that the encoding matches that supplied. This
 // is much slower, but verifies that we have the correct types defined
-function checkInstance (created: Codec, value: Uint8Array): void {
+function checkInstance (created: Codec, matcher: Uint8Array): void {
   const u8a = created.toU8a();
   const rawType = created.toRawType();
-
-  assert(
-    u8aEq(value, u8a) || (
-      // when length-prefixed from hex, just check the actual length
+  const isOk = (
+    // full match, all ok
+    u8aEq(u8a, matcher) ||
+    (
+      // on a length-prefixed type, just check the actual length
       ['Bytes', 'Text', 'Type'].includes(rawType) &&
-      value.length === (created as Bytes).length
-    ),
-    () => `${rawType}:: Decoded input doesn't match input, received ${u8aToHex(value, 512)} (${value.length} bytes), created ${u8aToHex(u8a, 512)} (${u8a.length} bytes)`
+      matcher.length === (created as Bytes).length
+    ) ||
+    (
+      // when the created is empty and matcher is also empty, let it slide...
+      created.isEmpty &&
+      matcher.every((v) => !v)
+    )
   );
+
+  assert(isOk, () => `${rawType}:: Decoded input doesn't match input, received ${u8aToHex(matcher, 512)} (${matcher.length} bytes), created ${u8aToHex(u8a, 512)} (${u8a.length} bytes)`);
 }
 
 function checkPedantic (created: Codec, [value]: unknown[], isPedantic = false): void {